HEAP-0008: renamed old test file; added two parameterized tests to new pytest file; added repr method to class
